move `_get_params_for_weight_decay_optimization` to apex.transformer.testing
