+++ tests/L0/run_transformer/run_bert_minimal_test.py
@@ -8,13 +8,13 @@ from apex.transformer.pipeline_parallel.utils import setup_microbatch_calculator
 from apex.transformer.pipeline_parallel.utils import average_losses_across_data_parallel_group
 from apex.transformer.pipeline_parallel.schedules import get_forward_backward_func
 from apex.transformer.pipeline_parallel.schedules.common import build_model
-from apex.transformer.pipeline_parallel.schedules.common import _get_params_for_weight_decay_optimization
 
 from apex.transformer.testing.standalone_bert import bert_model_provider
 from apex.transformer.testing import global_vars
 from apex.transformer.testing.commons import TEST_SUCCESS_MESSAGE
 from apex.transformer.testing.commons import initialize_distributed
 from apex.transformer.testing.commons import print_separator
+from apex.transformer.testing.commons import _get_params_for_weight_decay_optimization
 
 import warnings
 class DebugWarning(Warning):
