+++ tests/L0/run_transformer/run_megatron_gpt_pipeline.py
@@ -5,7 +5,6 @@ from typing import List
 import torch
 
 from apex.transformer import parallel_state
-from apex.transformer.pipeline_parallel.schedules.common import _get_params_for_weight_decay_optimization
 from apex.transformer.pipeline_parallel.schedules.common import build_model
 from apex.transformer.pipeline_parallel.schedules.fwd_bwd_pipelining_with_interleaving import _forward_backward_pipelining_with_interleaving
 from apex.transformer.pipeline_parallel.schedules.fwd_bwd_pipelining_without_interleaving import forward_backward_pipelining_without_interleaving
@@ -18,6 +17,7 @@ from apex.transformer.testing import global_vars
 from apex.transformer.testing.commons import TEST_SUCCESS_MESSAGE
 from apex.transformer.testing.commons import initialize_distributed
 from apex.transformer.testing.commons import print_separator
+from apex.transformer.testing.commons import _get_params_for_weight_decay_optimization
 from apex.transformer.testing.standalone_gpt import gpt_model_provider
 from apex.transformer.log_util import get_transformer_logger, set_logging_level
